refactor(plot): extract crosshair logic to separate files
Extracts `computeSnap`, `drawCanvas`, and `renderTooltipHTML` from the monolithic `Crosshair.tsx` component into a dedicated `crosshair/` subdirectory alongside the shared types. This resolves the "Large File Requires Refactoring" code health issue. Imports and functional bindings in the parent component have been adjusted accordingly while preserving original logic and performance characteristics.
